+++ NET1814_MilkShop.Repositories/Models/PaymentModels/PaymentDataModel.cs
@@ -3,5 +3,5 @@ namespace NET1814_MilkShop.Repositories.Models.PaymentModels;
 public class PaymentDataModel
 {
     public string CheckoutUrl { get; set; } = null!;
-    public int OrderCode { get; set; }
+    public long OrderCode { get; set; }
 }
